Project 3 - AZURE VECTOR SEARCH - OCR and RAG building/requirements.txt
quart
authlib
hypercorn
python-dotenv
openai
diskcache
azure-cognitiveservices-speech
aiohttp
azure-core
azure-search-document